LayoutTests/ietestcenter/Javascript/15.2.3.6-4-1-expected.txt

15.2.3.6-4-1 (TestCases/chapter15/15.2/15.2.3/15.2.3.6/15.2.3.6-4-1.js) Object.defineProperty throws TypeError when adding properties to non-extensible objects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
